% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_Regression.R
\name{coringTaxa}
\alias{coringTaxa}
\title{coringTaxa creates correlation dataframe for taxa}
\usage{
coringTaxa(microbeAbund)
}
\arguments{
\item{microbeAbund}{the taxa abundance dataframe (rownames sample names and colnames taxa Genus/species/family)}
}
\value{
A data frame of correlations between taxa
}
\description{
This function creates an output correlation data frame for all microbial taxa (or other organisms such as
viral or parasitic taxa)
}
\examples{
data(microbeAbund)
x <- coringTaxa(microbeAbund)

}
\keyword{correlation}
\keyword{taxa}
