% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CovFile}
\alias{CovFile}
\title{\code{mbQTL} "CovFile"}
\description{
The "CovFile" is the covariate file for linear regression option of taxa and snp association.
The covariance file is generated randomly by assigning sex and site of collection to
the samples.) rownames are covariate and colnames samples.
}
\keyword{data}
