% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowload_files.R
\name{downloadMbQTLFile}
\alias{downloadMbQTLFile}
\title{downloadMbQTLFile Read internal Miranda file}
\usage{
downloadMbQTLFile(urlf)
}
\arguments{
\item{urlf}{URL of the specific chosen file}
}
\value{
data.frame containing downloaded mbQTL file
}
\description{
Reads internal mbQTL file from extdata and returns it as a data.frame
}
\examples{
\donttest{
x <- downloadMbQTLFile(
       "https://zenodo.org/record/4615670/files/mbQTL.txt.gz"
     )
}
}
