% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowload_files.R
\name{importMbQTLFile}
\alias{importMbQTLFile}
\title{import Read internal Miranda file}
\usage{
importMbQTLFile(fn)
}
\arguments{
\item{fn}{filename}
}
\value{
data.frame containing Miranda data
}
\description{
Reads internal Miranda file from extdata and returns it as a data.frame
}
\examples{
\donttest{
x <- importMbQTLFile("Mouse_miRanda.txt")
}
}
