% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mbQTL-package}
\alias{mbQTL}
\alias{mbQTL-package}
\title{title \code{mbQTL} is a package for microbial QTL/GWAS Analysis}
\description{
This package provides statistical methods for identifying significant relationships between
microbial/taxa and genetic SNP signatures. We use three models 1) linear regression between
all taxa-snp. Main function is \code{linearTaxaSnp()}. 2) Correlation analysis between taxa-snp
across all taxa and snps. Main function is \code{taxaSnpCor()} and 3) Logistic regression analysis
between each taxa and each snp simultaneously or for a specific cases. Main function is
\code{logRegSnpsTaxa()}.
}
\seealso{
The package vignette can be accessed with \code{vignette("mbQTL")}.
}
\author{
\strong{Maintainer}: Mercedeh Movassagh \email{mercedeh.movassagh@yale.edu} (\href{https://orcid.org/0000-0001-7690-0230}{ORCID})

Authors:
\itemize{
  \item Steven Schiff
  \item Joseph N Paulson
}

}
