% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metagenomeSeqObj}
\alias{metagenomeSeqObj}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\keyword{MRexperiment}
\keyword{RDS}
\keyword{This}
\keyword{data}
\keyword{either}
\keyword{file}
\keyword{is}
\keyword{it}
\keyword{microbeAbund}
\keyword{or}
\keyword{the}
\keyword{this}
