% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microbeAbund}
\alias{microbeAbund}
\title{\code{mbQTL} "microbiomeAbund" File}
\description{
This is the microbial Abudnance file generated from 16s it is either this file or
the "metaGenomeSeqObj".The "microbiomeAbund" file is a randomly generated file
format for total microbe presence (number of reads)(parasite/viral transcripts)
for specific species.This could be generated from select taxa results from
\code{biom()} or \code{MRexperiment} objects as long as the samples are colnames and taxa
are rownames.
}
\keyword{data}
