% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logisitic_Regression.R
\name{prepareCorData}
\alias{prepareCorData}
\title{prepareCorData prpares and joins snp-taxa and taxa-taxa correlation file.}
\usage{
prepareCorData(microbeAbund, SnpFile, cutoff = NULL, selectmicrobe = NULL)
}
\arguments{
\item{microbeAbund}{the taxa abundance dataframe (rownames sample names and colnames taxa Genus/species/family)}

\item{SnpFile}{the snp dataframe (values 0,1,2 indicating zygosity), rownames sample names and colnames snp names.}

\item{cutoff}{default is NULL, hence anything above cutoff is considered positive, the cutoff at which the
specific or all taxa are considered positive for the pathogen (1 indicates positive and 0 negative).}

\item{selectmicrobe}{default is and all taxa are considered at the same time, if the user is interested in a
specific pathogen use name of the pathogen for example "Haemophilus".}
}
\value{
A data frame which is a result of Logistic regression  products of individual snp, taxa relationships,
with P values and P value corrected values.
}
\description{
This function creates a dataframe output produces a formatted dataframe prepared.
}
\examples{
\dontrun{
x <- prepareCorData(microbeAbund, SnpFile, cutoff = NULL, selectmicrobe = NULL)
}

}
\keyword{logitPlotDataframe}
