% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SnpFile}
\alias{SnpFile}
\title{\code{mbQTL} "SnpFile"}
\description{
The "SnpFile" is randomly generated from GATK (Van der Auwera GA & O'Connor BD. (2020).
Genomics in the Cloud: Using Docker, GATK, and WDL in Terra (1st Edition).
O'Reilly Media) snp calls followed by plink (Purcell S, et al. (2007) PLINK: a
toolset for whole-genome association and population-based linkage analysis. American
Journal of Human Genetics) processing and it needs to be in (0,1,2) format representing
the zygosity of the snps.
}
\keyword{data}
