% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_Regression.R
\name{TaxaSnpCor}
\alias{TaxaSnpCor}
\title{TaxaSnpCor for estimation of the rho value between snp, taxa correlations across datasets}
\usage{
TaxaSnpCor(for_all_rsids, correlationMicrobes, probs = NULL)
}
\arguments{
\item{for_all_rsids}{A dataframe result of correlation analysis between the snps and taxa dataframe,
an output of all_to_all_product() function.}

\item{correlationMicrobes}{A dataframe of correlation between coringTaxa() function.}

\item{probs}{Default is NULL if other that all rho values are wanted the value can be subseted using c(x,y).}
}
\value{
A data frame of correlations between taxa
}
\description{
This function produces a log heatmap +1 of the correlation rho values across snp, taxa dataframe.
}
\examples{
x <-TaxaSnpCor(for_all_rsids,correlationMicrobes)
}
\keyword{estimation}
\keyword{rho}
