% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CovFile}
\alias{CovFile}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\keyword{This}
\keyword{and}
\keyword{association}
\keyword{covariate}
\keyword{data}
\keyword{file}
\keyword{for}
\keyword{is}
\keyword{linear}
\keyword{of}
\keyword{option}
\keyword{regression}
\keyword{snp}
\keyword{taxa}
\keyword{the}
