% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_regression.R
\name{histPvalueLm}
\alias{histPvalueLm}
\title{histPvalueLm a histogram of Taxa and SNP linear regression analysis.
This function creates a histogram object of all SNPs with all taxa Linear regression analysis p values.}
\usage{
histPvalueLm(LinearAnalysisTaxaSNP)
}
\arguments{
\item{LinearAnalysisTaxaSNP}{the data frame result created from the linearTaxaSnp() function.}
}
\value{
A histogram object of p values observed from taxa and SNP Linear Regression analysis.
}
\description{
histPvalueLm a histogram of Taxa and SNP linear regression analysis.
This function creates a histogram object of all SNPs with all taxa Linear regression analysis p values.
}
\examples{
x <- qqPlotLm(microbeAbund, SnpFile, Covariate = CovFile)

}
\keyword{histogram}
\keyword{linear_regression}
\keyword{plot}
\keyword{snp}
\keyword{taxa}
