% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metagenomeSeqObj}
\alias{metagenomeSeqObj}
\title{\code{mbQTL} "metagenomeSeqObj"

"MetagenomSeqObj" is an \code{MRexperiment} object format of the "microbeAbund" file.}
\description{
\code{mbQTL} "metagenomeSeqObj"

"MetagenomSeqObj" is an \code{MRexperiment} object format of the "microbeAbund" file.
}
\keyword{data}
