% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microbeAbund}
\alias{microbeAbund}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\keyword{16s}
\keyword{Abudnance}
\keyword{This}
\keyword{data}
\keyword{either}
\keyword{file}
\keyword{from}
\keyword{generated}
\keyword{is}
\keyword{it}
\keyword{metaGenomeSeqObj}
\keyword{microbial}
\keyword{or}
\keyword{the}
\keyword{this}
