% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_Regression.R
\name{RegSnp}
\alias{RegSnp}
\title{RegSnp creates a dataframe of parsed long snp files}
\usage{
RegSnp(SnpFile, microbeAbund)
}
\arguments{
\item{SnpFile}{the snp file (rownames is sample number and colnames is the snps)}
}
\value{
A long parsed datframe of snps
}
\description{
This internal function takes the orginal snp dataframe and retruns a long parsed snp dataframe
}
\examples{
x <-RegSnp(SnpFile)
}
\keyword{long}
\keyword{parsed}
\keyword{snp}
