% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SnpFile}
\alias{SnpFile}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\keyword{/or}
\keyword{Snp}
\keyword{This}
\keyword{data}
\keyword{file}
\keyword{from}
\keyword{genome}
\keyword{host}
\keyword{is}
\keyword{pathogen}
\keyword{the}
